Isolating solve challenge links list to its own component
