+++ web/src/pages/ChallengeDetailsPage.js
@@ -3,8 +3,7 @@ import { useSelector, useDispatch } from "react-redux";
 import { Helmet } from "react-helmet";
 import { css } from "@emotion/core";
 import PropTypes from "prop-types";
-import { Page, Grid, Dimmer, Button, Icon } from "tabler-react";
-import { Terminal } from "react-window-ui";
+import { Page, Grid, Dimmer, Button } from "tabler-react";
 import siteMetaData from "../constants/metadata";
 import {
   fetchChallengeDetail as fetchChallengeDetailAction,
@@ -16,46 +15,12 @@ import ChallengeBuyButton from "../components/challenges/ChallengeBuyButton";
 import ChallengeValidateButton from "../components/challenges/ChallengeValidateButton";
 import ChallengeCloseButton from "../components/challenges/ChallengeCloseButton";
 import ValidationsList from "../components/challenges/ValidationsList";
+import ChallengeSolveInstances from "../components/challenges/ChallengeSolveInstances";
 import { CLEAN_CHALLENGE_DETAIL } from "../constants/actionTypes";
 
 const paragraph = css`
   margin-top: 0.5rem;
 `;
-
-const ChallengeSolveInstances = ({ instances }) => {
-  const terminal = css`
-    margin-bottom: 1rem;
-
-    a {
-      color: #fff;
-      margin-right: 1.5rem;
-    }
-
-    span {
-      color: #16b279;
-    }
-  `;
-
-  return (
-    <Terminal
-      minHeight="2rem"
-      boxShadow="0px 2px 15px -8px rgba(0,0,0,0.41)"
-      css={terminal}
-    >
-      {instances.map(item => (
-        <div key={item.id}>
-          <a href={item.nginx_url} target="_blank" rel="noreferrer">
-            {item.nginx_url}
-          </a>
-          <span>
-            <Icon name="check-circle" />
-            {item.status}
-          </span>
-        </div>
-      ))}
-    </Terminal>
-  );
-};
 
 const ChallengeDetailsPage = props => {
   const dispatch = useDispatch();
